% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-agreement.R
\docType{data}
\name{agreement_data}
\alias{agreement_data}
\title{Agreement data for N=20 athletes}
\format{
A data frame with 20 rows and 6 variables:
\describe{
\item{Athlete}{Name of the athlete}
\item{True_score}{Numeric value indicating true jump height in cm units}
\item{Criterion_score.trial1}{Criterion measure of the true score - trial 1}
\item{Criterion_score.trial2}{Criterion measure of the true score - trial 2}
\item{Practical_score.trial1}{Practical measure of the true score - trial 1}
\item{Practical_score.trial2}{Practical measure of the true score - trial 2}
}
}
\usage{
data(agreement_data)
}
\description{
Data for true, criterion and practical vertical jump measures. Criterion and practical
were measured twice (trial1 and trial2), assuming equal true score
}
\keyword{datasets}
