% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-fit.R
\name{baseline_model}
\alias{baseline_model}
\title{Baseline model}
\usage{
baseline_model(
  predictors,
  outcome,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{predictors}{Data frame}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
\code{model} object
}
\description{
This model return the mean of the \code{outcome} as a prediction
}
\examples{
baseline_model(
  predictors = iris[2:3],
  outcome = iris[[1]]
)
}
