% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_arguments.R
\name{basic_arguments}
\alias{basic_arguments}
\title{Basic arguments}
\arguments{
\item{group_a}{Numeric vector. This groups represents baseline/control or Pre-test in the paired design}

\item{group_b}{Numeric vector}

\item{paired}{Paired groups? Default is \code{FALSE}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}
}
\description{
Basic arguments
}
