% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_press_data.R
\docType{data}
\name{bench_press_data}
\alias{bench_press_data}
\title{Bench press data}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{Athlete}{Name of the athlete}
  \item{Pre-test}{Bench press 1RM Pre-test results in kg}
  \item{Post-test}{Bench press 1RM Post-test results in kg}
  \item{Change}{Difference between Post-test and Pre-test}
  \item{Magnitude}{Factor indicating magnitude of the change based on
                   the SESOI band (+/- 5kg). Three levels: Lower, Equivalent,
                   and Higher.}
}
}
\usage{
data(bench_press_data)
}
\description{
Effects of training intervention using N=20 subject on bench press 1RM
}
\keyword{datasets}
