% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats-package.R
\docType{package}
\name{bmbstats-package}
\alias{bmbstats-package}
\alias{_PACKAGE}
\title{bmbstats: Bootstrap Magnitude-based Statistics}
\description{
Set of functions for analysis and graphing of common sport science problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mladenjovanovic/bmbstats}
  \item \url{https://github.com/mladenjovanovic/bmbstats-book}
  \item Report bugs at \url{https://github.com/mladenjovanovic/bmbstats/issues}
}

}
\author{
\strong{Maintainer}: Mladen Jovanović \email{coach.mladen.jovanovic@gmail.com}

}
\keyword{internal}
