% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats.R
\name{bmbstats}
\alias{bmbstats}
\title{Bootstrap Magnitude-based generic function}
\usage{
bmbstats(
  data,
  SESOI_lower_function = function(data, na.rm, init_boot) {     return(0) },
  SESOI_upper_function = function(data, na.rm, init_boot) {     return(0) },
  estimator_function = function(data, SESOI_lower, SESOI_upper, na.rm, init_boot) {    
    return(0) },
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{SESOI_lower_function}{Function to estimate SESOI_lower within bootstrap loop.
Default functions is \code{function(data, na.rm, init_boot) {return(0)}}}

\item{SESOI_upper_function}{Function to estimate SESOI_upper within bootstrap loop.
Default functions is \code{function(data, na.rm, init_boot) {return(0)}}}

\item{estimator_function}{Function to be used within the bootstrap loop to provide a list
or named numeric vector of parameters.
Default is \code{function(data, SESOI_lower, SESOI_upper, na.rm, init_boot){return(0)}}}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata} to setup bootstrap.}

\item{na.rm}{Should missing values be removed? Default is \code{FALSE}.
This is forwarded to \code{estimator_function}, \code{SESOI_lower_function}, and \code{SESOI_upper_function}}
}
\value{
A \code{bmbstats} object
}
\description{
This is the "core" function of the \code{bmbstats} package.
It performs bootstrap on the \code{data} using provided \code{estimator_function},
\code{SESOI_lower_function}, and \code{SESOI_upper_function}. Used in other functions ("wrappers")
}
\examples{
bmbstats(iris,
  SESOI_lower_function = function(data, na.rm, init_boot) {
    sd(data$Sepal.Length) * -0.2
  },
  SESOI_upper_function = function(data, na.rm, init_boot) {
    sd(data$Sepal.Length) * 0.2
  },
  estimator_function = function(data, SESOI_lower, SESOI_upper, na.rm, init_boot) {
    list(mean = mean(data$Sepal.Length), SESOI_lower = SESOI_lower, SESOI_upper = SESOI_upper)
  },
  control = model_control(boot_type = "perc", boot_samples = 50)
)
}
