% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBI.R
\name{bootstrap_MBI}
\alias{bootstrap_MBI}
\title{Bootstrap magnitude based inference}
\usage{
bootstrap_MBI(bmbstats_object, estimator, SESOI_lower = 0, SESOI_upper = 0)
}
\arguments{
\item{bmbstats_object}{Object of class \code{bmbstats}}

\item{estimator}{Name of the estimator from the \code{bmbstats_object}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}
}
\value{
\code{bmbstats_MBI} object
}
\description{
Calculates proportion of the estimator bootstrap resample lower/equivalent/higher distribution compared to SESOI band
}
\examples{
mean_MBI <- bootstrap_MBI(
  describe_data(rnorm(10, 100, 20)),
  estimator = "mean",
  SESOI_lower = 95,
  SESOI_upper = 105
)
mean_MBI
plot(mean_MBI, control = plot_control(points_jitter = FALSE))
}
