% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MET.R
\name{bootstrap_MET}
\alias{bootstrap_MET}
\title{Bootstrap minimum effect tests (METs)}
\usage{
bootstrap_MET(
  bmbstats_object,
  estimator,
  SESOI_lower = 0,
  SESOI_upper = 0,
  alpha = 0.05
)
}
\arguments{
\item{bmbstats_object}{Object of class \code{bmbstats}}

\item{estimator}{Name of the estimator from the \code{bmbstats_object}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{alpha}{Numeric value for the alpha (Type I error). Default is 0.05}
}
\value{
\code{bmbstats_MET} object
}
\description{
Bootstrap minimum effect tests (METs)
}
\examples{
mean_MET <- bootstrap_MET(
  describe_data(rnorm(10, 100, 20)),
  estimator = "mean",
  SESOI_lower = 95,
  SESOI_upper = 105,
  alpha = 0.05
)
mean_MET
plot(mean_MET)
}
