% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHST.R
\name{bootstrap_NHST}
\alias{bootstrap_NHST}
\title{Bootstrap Null Hypothesis Significance Test}
\usage{
bootstrap_NHST(
  bmbstats_object,
  estimator,
  null_hypothesis = 0,
  test = "two.sided"
)
}
\arguments{
\item{bmbstats_object}{Object of class \strong{bmbstats}}

\item{estimator}{Name of the estimator from the \code{bmbstats_object}}

\item{null_hypothesis}{Numeric scalar indicating null hypothesis value for \code{estimator}. Default is 0}

\item{test}{Character string specifying the type of NHST, must be one of "two.sided" (default), "greater" or "less".
Default is "two.sided"}
}
\value{
\code{bmbstats_NHST} object
}
\description{
Bootstrap Null Hypothesis Significance Test
}
\examples{
mean_NHST <- bootstrap_NHST(
  describe_data(rnorm(100)),
  "mean"
)
mean_NHST
plot(mean_NHST)
}
