% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Cohen's d for paired (dependent) and unpaired (independent) samples}
\usage{
cohens_d(group_a, group_b, paired = FALSE, na.rm = FALSE)
}
\arguments{
\item{group_a}{Numeric vector. This groups represents baseline/control or Pre-test in the paired design}

\item{group_b}{Numeric vector}

\item{paired}{Paired groups? Default is \code{FALSE}}

\item{na.rm}{Should missing values be removed? Default is \code{FALSE}. With \code{paired=TRUE}, NA removal is \emph{NOT} pairwise
and involves NA removal from the difference score and \code{group_a}}
}
\value{
Cohen's d
}
\description{
Cohen's d for paired (dependent) and unpaired (independent) samples
}
\examples{
cohens_d(rnorm(100), rnorm(100))
}
