% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dependent_groups.R
\name{compare_dependent_groups}
\alias{compare_dependent_groups}
\title{Compare two dependent groups}
\usage{
compare_dependent_groups(
  pre,
  post,
  SESOI_lower = SESOI_lower_dependent_func,
  SESOI_upper = SESOI_upper_dependent_func,
  estimator_function = dependent_groups_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{pre}{Numeric vector}

\item{post}{Numeric vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{estimator_function}{Function that takes \code{pre}, \code{post}, and \code{na.rm} parameters
and return named numeric vector with parameter estimates. Default is \code{\link{dependent_groups_estimators}}.
User can write their own function with needed estimators for which bootstrap confidence intervals
are needed}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
\code{compare_dependent_groups} provides numerous descriptive estimators and bootstrap confidence intervals for
comparison between paired \code{pre} and \code{post}. \code{compare_dependent_groups} function is a "wrapper"
for \code{\link{bmbstats}} function.
}
\examples{
data("bench_press_data")
pre <- bench_press_data$`Pre-test`
post <- bench_press_data$`Post-test`
compare_dependent_groups(pre, post, SESOI_lower = -5, SESOI_upper = 5)
}
