% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_independent_groups.R
\name{compare_independent_groups}
\alias{compare_independent_groups}
\title{Compare two independent groups}
\usage{
compare_independent_groups(
  group_a,
  group_b,
  SESOI_lower = SESOI_lower_independent_func,
  SESOI_upper = SESOI_upper_independent_func,
  estimator_function = independent_groups_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, observed variable, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, predicted variable, Post-test in the paired design, or "criterion" measure}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{estimator_function}{Function that takes \code{group_a}, \code{group_b}, and \code{na.rm} parameters
and return named numeric vector with parameter estimates. Default is \code{\link{independent_groups_estimators}}.
User can write their own function with needed estimators for which bootstrap confidence intervals
are needed}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
\code{compare_independent_groups} provides numerous descriptive estimators and bootstrap confidence intervals for
comparison between \code{group_a} and \code{group_b}. \code{compare_independent_groups} function is a "wrapper"
for \code{\link{bmbstats}} function.
}
\examples{
data("height_data")
group_a <- height_data$Height[height_data$Gender == "Female"]
group_b <- height_data$Height[height_data$Gender == "Male"]
compare_independent_groups(group_a, group_b, SESOI_lower = -2.5, SESOI_upper = 2.5)
}
