% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-functions.R
\name{cost_functions}
\alias{cost_functions}
\alias{cost_MBE}
\alias{cost_MAE}
\alias{cost_MSE}
\alias{cost_RMSE}
\alias{cost_SSE}
\alias{cost_MinErr}
\alias{cost_MaxErr}
\alias{cost_MaxAbsErr}
\alias{cost_MaxSqErr}
\alias{cost_R_squared}
\alias{cost_SESOItoRMSE}
\alias{cost_PPER}
\alias{cost_MHE}
\alias{cost_RMHE}
\title{Cost functions}
\usage{
cost_MBE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MAE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MSE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_RMSE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_SSE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MinErr(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MaxErr(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MaxAbsErr(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MaxSqErr(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_R_squared(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_SESOItoRMSE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_PPER(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_MHE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)

cost_RMHE(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  negative_weight = 1,
  positive_weight = 1,
  na.rm = FALSE
)
}
\arguments{
\item{observed}{Numeric vector}

\item{predicted}{Numeric vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{negative_weight}{How should negative residuals be weighted? Default is 1}

\item{positive_weight}{How should positive residuals be weighted? Default is 1}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
This method uses algebraic method assuming normal distribution of the residuals.
This is done by using \code{\link[stats]{sd}} rather than RSE from
\code{\link[stats]{lm}} model.
}
\examples{
data("yoyo_mas_data")

model <- lm(MAS ~ YoYoIR1, yoyo_mas_data)

observed <- yoyo_mas_data$MAS
predicted <- predict(model)

SESOI_lower <- -0.5
SESOI_upper <- 0.5

# Mean Squared Error
cost_MSE(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)

# Mean Absolute Error
cost_MAE(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)

# Root Mean Squared Error
cost_RMSE(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)

# Bias
cost_MBE(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)

# Sum of Squared Errors
cost_SSE(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)

# Proportion of Practically Equivalent Residuals
cost_PPER(
  observed = observed,
  predicted = predicted,
  SESOI_lower = SESOI_lower,
  SESOI_upper = SESOI_upper
)
}
