% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model-fit.R
\name{cv_model}
\alias{cv_model}
\alias{cv_model.default}
\alias{cv_model.data.frame}
\alias{cv_model.matrix}
\alias{cv_model.formula}
\alias{cv_model.recipe}
\title{Fit a \code{cv_model}}
\usage{
cv_model(x, ...)

\method{cv_model}{default}(x, ...)

\method{cv_model}{data.frame}(x, y, ...)

\method{cv_model}{matrix}(x, y, ...)

\method{cv_model}{formula}(formula, data, ...)

\method{cv_model}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{See Details}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{bmbstats_cv_model} object.
}
\description{
\code{cv_model()} fits a model.
}
\details{
Will be expanded
}
\examples{
m1 <- cv_model(
  Sepal.Length~. -Species,
  iris
)
predict(m1, new_data = iris)
}
