% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-data.R
\name{data_estimators}
\alias{data_estimators}
\title{Data estimators}
\usage{
data_estimators(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector with estimators
}
\description{
Function that provides a list of estimators. User is free to define his own list of estimators.
Used in \code{\link{describe_data}}
}
\examples{
data_estimators(iris$Sepal.Width)
}
