% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-data.R
\name{data_estimators_robust}
\alias{data_estimators_robust}
\title{Data estimators (robust)}
\usage{
data_estimators_robust(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector with estimators
}
\description{
Function that provides a list of robust estimators. User is free to define his own list of estimators.
Used in \code{\link{describe_data}}
}
\examples{
data_estimators_robust(iris$Sepal.Width)
}
