% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-data.R
\name{data_estimators_simple}
\alias{data_estimators_simple}
\title{Data estimators (simple)}
\usage{
data_estimators_simple(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector with estimators \code{\link[base]{mean}}, and \code{\link[stats]{sd}}
}
\description{
Function that provides a simple list of estimators (mean and SD).
User is free to define his own list of estimators. Used in \code{\link{describe_data}}
}
\examples{
data_estimators_simple(iris$Sepal.Width)
}
