% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{density_mode}
\alias{density_mode}
\title{Finding Modes Using Kernel Density Estimates}
\source{
Website page \url{https://rmflight.github.io/post/finding-modes-using-kernel-density-estimates/}
}
\usage{
density_mode(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Mode estimate
}
\description{
Finding Modes Using Kernel Density Estimates
}
\examples{
density_mode(rnorm(100))
}
