% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dependent_groups.R
\name{dependent_groups_estimators}
\alias{dependent_groups_estimators}
\title{Dependent groups estimators}
\usage{
dependent_groups_estimators(
  pre,
  post,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{pre}{Numeric vector}

\item{post}{Numeric vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector with estimators
}
\description{
Function that provides a list of estimators. User is free to define his own list of estimators.
Used in \code{\link{compare_dependent_groups}}
}
\examples{
data("bench_press_data")
pre <- bench_press_data$`Pre-test`
post <- bench_press_data$`Post-test`
dependent_groups_estimators(pre, post, SESOI_lower = -5, SESOI_upper = 5)
}
