% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-data.R
\name{describe_data}
\alias{describe_data}
\title{Describe data}
\usage{
describe_data(
  x,
  estimator_function = data_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{x}{Numeric vector}

\item{estimator_function}{Function that takes \code{x} and \code{na.rm} parameters and return named
numeric vector with parameter estimates. Default is \code{\link{data_estimators}}. Other functions
that could be used are \code{\link{data_estimators_simple}} and \code{\link{data_estimators_robust}}.
User can write their own function with needed estimators for which bootstrap confidence intervals
are needed}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
\code{describe_data} provides numerous descriptive estimators and bootstrap confidence intervals for
variable \code{x}. \code{describe_data} function is a "wrapper" for \code{\link{bmbstats}} function.
}
\examples{
desc1 <- describe_data(
  iris$Sepal.Width,
  estimator_function = data_estimators_simple,
  control = model_control(confidence = 0.90, boot_samples = 1000)
)

desc1
plot(desc1)
}
