% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_relationship.R
\name{describe_relationship}
\alias{describe_relationship}
\title{Describe relationship between two dependent groups}
\usage{
describe_relationship(
  predictor,
  outcome,
  SESOI_lower = SESOI_lower_dependent_func,
  SESOI_upper = SESOI_upper_dependent_func,
  estimator_function = relationship_lm_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{predictor}{Numeric vector}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{estimator_function}{Function that takes \code{predictor}, \code{outcome}, and \code{na.rm} parameters
and return named numeric vector with parameter estimates. Default is \code{\link{relationship_lm_estimators}}.
User can write their own function with needed estimators for which bootstrap confidence intervals
are needed}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
\code{describe_relationship} provides numerous descriptive estimators and bootstrap confidence intervals for
relationship between paired \code{predictor} and \code{outcome} groups. \code{describe_relationship} function is a "wrapper"
for \code{\link{bmbstats}} function.
}
\examples{
data("yoyo_mas_data")
predictor <- yoyo_mas_data$YoYoIR1
outcome <- yoyo_mas_data$MAS

describe_relationship(predictor, outcome, SESOI_lower = -0.5, SESOI_upper = 0.5)
}
