% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_num}
\alias{func_num}
\title{Function or numeric}
\usage{
func_num(x, ...)
}
\arguments{
\item{x}{Function or numeric}

\item{...}{Forwarded to \code{x} if function}
}
\value{
If function \code{x{...}} is returned. If numeric \code{x} is returned
}
\description{
Check if the parameter \code{x} is a function, or numeric. If it is a function,
then parameters \code{...} are forwarded to \code{x}. If it is a numeric, then
the \code{x} is returned
}
