% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-fit.R
\name{generic_predict}
\alias{generic_predict}
\title{Predict using generic function}
\usage{
generic_predict(
  model,
  predictors,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{predictors}{Data frame}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
Predicts using \code{\link[stats]{predict}} function and supplied \code{model}
}
\examples{
m1 <- lm_model(
  predictors = iris[2:3],
  outcome = iris[[1]]
)

generic_predict(m1)
}
