% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats.R
\name{get_bootstrap_ci}
\alias{get_bootstrap_ci}
\title{Gets bootstrap confidence intervals}
\usage{
get_bootstrap_ci(boot_object, control = model_control())
}
\arguments{
\item{boot_object}{Bootstrap object returned from \code{link[boot]{boot}} function}

\item{control}{Control object returned from \code{\link{model_control}} function.}
}
\value{
Data frame
}
\description{
Function to extract confidence intervals from bootstrap object
}
