% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_data.R
\docType{data}
\name{height_data}
\alias{height_data}
\title{Height data for N=100 athletes}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Gender}{Factor indicating gender of the athletes. Two levels "Male" and "Female"}
\item{Height}{Numeric value indicating height in cm units}
}
}
\usage{
data(height_data)
}
\description{
Height data for N=50 males and N=50 females
}
\keyword{datasets}
