% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-fit.R
\name{lm_model}
\alias{lm_model}
\title{Simple linear regression model}
\usage{
lm_model(
  predictors,
  outcome,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{predictors}{Data frame}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}

\item{...}{Extra arguments forwarded to \code{\link[stats]{lm}} function}
}
\value{
\code{model} object
}
\description{
This model uses all \code{predictors} to model the \code{outcome} with
\code{\link[stats]{lm}} function
}
\examples{
lm_model(
  predictors = iris[2:3],
  outcome = iris[[1]]
)
}
