% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb-proportions.R
\name{mb_proportions}
\alias{mb_proportions}
\title{Magnitude-based Proportions for Two Groups}
\usage{
mb_proportions(
  group_a,
  group_b,
  paired = FALSE,
  SESOI_lower = 0,
  SESOI_upper = 0,
  method = "algebraic",
  na.rm = FALSE,
  use_normal_distribution = FALSE
)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, observed variable, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, predicted variable, Post-test in the paired design, or "criterion" measure}

\item{paired}{Paired groups? Default is \code{FALSE}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{method}{Select "brute-force" or "algebraic" method to calculate proportions. Default is "algebraic"}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}

\item{use_normal_distribution}{When estimating proportions algebraically, should normal or t-distribution
be used. Default is \code{FALSE} (normal t-distribution)}
}
\value{
Data frame with lower, equivalent and higher columns
}
\description{
\code{mb_proportions} calculates lower/equivalent/higher difference
proportions between two groups
}
\examples{
mb_proportions(rnorm(100), rnorm(100))
}
