% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_proportions.R
\name{mb_proportions_dependent}
\alias{mb_proportions_dependent}
\title{Magnitude-based Proportions for Two Dependent Groups}
\usage{
mb_proportions_dependent(
  group_a,
  group_b,
  SESOI_lower = 0,
  SESOI_upper = 0,
  method = "algebraic",
  na.rm = FALSE,
  use_normal_distribution = TRUE
)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, Post-test in the paired design, or "criterion" measure}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{method}{Select "brute-force" or "algebraic" method to calculate proportions. Default is "algebraic"}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}

\item{use_normal_distribution}{When estimating proportions algebraically, should normal or t-distribution
be used. Default is TRUE (normal distribution)}
}
\value{
Data frame with lower, equivalent and higher columns
}
\description{
\code{mb_proportions_dependent} calculates lower/equivalent/higher
   difference proportions between two dependent groups
}
