% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_sd_v}
\alias{mean_sd_v}
\title{Define  \code{\link[ggplot2]{ggplot}} summary function to return mean +/- SD VERTICAL}
\usage{
mean_sd_v(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Data frame with xmin, x, and xmax summaries
}
\description{
Define  \code{\link[ggplot2]{ggplot}} summary function to return mean +/- SD VERTICAL
}
\examples{
mean_sd_v(rnorm(100))
}
