% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations-MBI.R
\name{observations_MBI}
\alias{observations_MBI}
\title{Magnitude-based inference using observations and known measurement error}
\usage{
observations_MBI(
  observations,
  observations_label = NULL,
  measurement_error = 0,
  df = Inf,
  SESOI_lower = 0,
  SESOI_upper = 0,
  confidence = 0.95
)
}
\arguments{
\item{observations}{Numeric vector of observed scores}

\item{observations_label}{Vector used for labeling observations. Default is \code{NULL}}

\item{measurement_error}{Estimated SD of the random error. Default is 0}

\item{df}{Degrees of freedom from the reliability study where \code{measurement_error} is
estimated. Default is \code{Inf}, which implies the use of normal distribution}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{confidence}{Default is 0.95}
}
\description{
Magnitude-based inference using observations and known measurement error
}
\examples{
data("bench_press_data")
obs_MBI <- observations_MBI(
  bench_press_data$Change,
  SESOI_lower = -5,
  SESOI_upper = 5,
  measurement_error = 2.5,
  confidence = 0.9
)

obs_MBI
plot(obs_MBI)
}
