% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations-MET.R
\name{observations_MET}
\alias{observations_MET}
\title{Minimum-effect tests using observations and known measurement error}
\usage{
observations_MET(
  observations,
  observations_label = NULL,
  measurement_error = 0,
  df = Inf,
  SESOI_lower = 0,
  SESOI_upper = 0,
  alpha = 0.05,
  confidence = 0.95
)
}
\arguments{
\item{observations}{Numeric vector of observed scores}

\item{observations_label}{Vector used for labeling observations. Default is \code{NULL}}

\item{measurement_error}{Estimated SD of the random error. Default is 0}

\item{df}{Degrees of freedom from the reliability study where \code{measurement_error} is
estimated. Default is \code{Inf}, which implies the use of normal distribution}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{alpha}{Numeric value for the alpha (Type I error). Default is 0.05}

\item{confidence}{Default is 0.95}
}
\description{
Minimum-effect tests using observations and known measurement error
}
\examples{
data("bench_press_data")
obs_MET <- observations_MET(
  bench_press_data$Change,
  SESOI_lower = -5,
  SESOI_upper = 5,
  measurement_error = 2.5,
  alpha = 0.05
)

obs_MET
plot(obs_MET)
}
