% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perfect_rnorm}
\alias{perfect_rnorm}
\title{Perfect Normal Distribution
Generates perfect normally distributed numbers (used for plotting)}
\usage{
perfect_rnorm(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{Number of samples}

\item{mean}{Mean of the distribution. Default is 0}

\item{sd}{SD of the distribution. Default is 1}
}
\value{
Numeric vector, length of \code{n}
}
\description{
Perfect Normal Distribution
Generates perfect normally distributed numbers (used for plotting)
}
\examples{
perfect_rnorm(10)
}
