% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model-fit.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance metrics}
\usage{
performance_metrics(
  observed,
  predicted,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{observed}{Numeric vector}

\item{predicted}{Numeric vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
Returns named vector of the most common performance metrics
}
\examples{
m1 <- lm_model(
  predictors = iris[2:3],
  outcome = iris[[1]]
)

predicted <- generic_predict(m1)

performance_metrics(
  observed = iris[[1]],
  predicted = predicted
)
}
