% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats-S3.R
\name{plot.bmbstats}
\alias{plot.bmbstats}
\title{S3 method for plotting \code{\link{bmbstats}} results}
\usage{
\method{plot}{bmbstats}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats}}

\item{...}{Extra arguments. Use \code{\link{plot_control}} to control plotting style}
}
\description{
This function plots the bootstrap distribution
}
\examples{
x <- bmbstats(iris,
  SESOI_lower_function = function(data, na.rm, init_boot) {
    sd(data$Sepal.Length) * -0.2
  },
  SESOI_upper_function = function(data, na.rm, init_boot) {
    sd(data$Sepal.Length) * 0.2
  },
  estimator_function = function(data, SESOI_lower, SESOI_upper, na.rm, init_boot) {
    list(mean = mean(data$Sepal.Length), SESOI_lower = SESOI_lower, SESOI_upper = SESOI_upper)
  },
  control = model_control(boot_type = "perc", boot_samples = 50)
)
plot(x)
}
