% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MET-S3.R
\name{plot.bmbstats_MET}
\alias{plot.bmbstats_MET}
\title{S3 method for plotting \code{\link{bootstrap_MET}} results}
\usage{
\method{plot}{bmbstats_MET}(x, type = "CI", ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_MET}}

\item{type}{What test to plot? Options are "inferiority", "non-superiority",
"equivalence", "non-inferiority",  "superiority", "CI". Default is "CI"}

\item{...}{Extra arguments. Use \code{plot_control} for plotting options}
}
\description{
S3 method for plotting \code{\link{bootstrap_MET}} results
}
\examples{
mean_MET <- bootstrap_MET(
  describe_data(rnorm(10, 100, 20)),
  estimator = "mean",
  SESOI_lower = 95,
  SESOI_upper = 105,
  alpha = 0.05
)
mean_MET
plot(mean_MET)
}
