% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHST-S3.R
\name{plot.bmbstats_NHST}
\alias{plot.bmbstats_NHST}
\title{S3 method for plotting \code{\link{bootstrap_NHST}} results}
\usage{
\method{plot}{bmbstats_NHST}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_NHST}}

\item{...}{Extra arguments. Use \code{plot_control} for plotting options, particularly
\code{effect_colors} parameter}
}
\description{
S3 method for plotting \code{\link{bootstrap_NHST}} results
}
\examples{
mean_NHST <- bootstrap_NHST(
  describe_data(rnorm(100)),
  "mean"
)
mean_NHST
plot(mean_NHST)
}
