% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-analysis-S3.R
\name{plot.bmbstats_RCT_analysis}
\alias{plot.bmbstats_RCT_analysis}
\title{S3 method for plotting \code{\link{RCT_analysis}} results}
\usage{
\method{plot}{bmbstats_RCT_analysis}(x, type = "boot", ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_RCT_analysis}}

\item{type}{Type of plot. Options are "boot", "control-pre-post", "treatment-pre-post",
"control-change", "treatment-change", "change", "control-paired-change",
"treatment-paired-change", "change-distribution", "effect-distribution", and "treatment-responses".
Default is "boot"}

\item{...}{Extra arguments. Use \code{\link{plot_control}} to control plotting style}
}
\description{
S3 method for plotting \code{\link{RCT_analysis}} results
}
\examples{
set.seed(1666)

data("vertical_jump_data")

rct_model <- RCT_analysis(
  vertical_jump_data,
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control",
  pre_test = "Pre-test",
  post_test = "Post-test",
  control = model_control(boot_type = "perc", boot_samples = 500)
)

rct_model

plot(rct_model)
}
