% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-predict-S3.R
\name{plot.bmbstats_RCT_predict}
\alias{plot.bmbstats_RCT_predict}
\title{S3 method for plotting \code{\link{RCT_predict}} results}
\usage{
\method{plot}{bmbstats_RCT_predict}(x, type = "residuals", ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_RCT_predict}}

\item{type}{Type of plot. Options are "residuals", "prediction", "bias-variance", "pdp+ice", "counterfactual", and
"ice". Default is "residuals"}

\item{...}{Extra arguments. Use \code{control} argument and \code{\link{plot_control}} function to control plotting style.
\code{confidence} for setting confidence level, \code{metric} for selecting performance metric (default
is "RMSE") and \code{metric_CV} for selecting source of CV metrics (defaults is "testing.pooled"). For "pdp+ice",
use \code{predictor} to select predictor for plotting (default is \code{group} parameter from \code{\link{RCT_predict}}).}
}
\description{
S3 method for plotting \code{\link{RCT_predict}} results
}
\examples{
data("vertical_jump_data")

m1 <- cv_model(
  `Post-test` ~ `Pre-test` + Group,
  vertical_jump_data,
  control = model_control(
    cv_repeats = 10,
    cv_folds = 3,
    cv_strata = vertical_jump_data$Group
  )
)

m1_rct <- RCT_predict(
  m1,
  new_data = vertical_jump_data,
  outcome = "Post-test",
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control"
)

m1_rct
plot(m1_rct)
}
