% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{plot_control}
\alias{plot_control}
\title{Plot control constructor}
\usage{
plot_control(
  group_colors = c("#5DA5DA", "#F15854", "#4D4D4D", "#60BD68", "#FAA43A", "#F17CB0",
    "#B276B2", "#DECF3F"),
  effect_colors = c(Lower = "#F158546A", Equivalent = "#4D4D4D6A", Higher =
    "#60BD686A"),
  cloud_scale = 0.5,
  cloud_alpha = 0.6,
  cloud_color = NA,
  cloud_quantile_lines = TRUE,
  cloud_quantile_lines_size = 0.4,
  cloud_quantile_lines_color = "white",
  cloud_quantile_lines_adjust = FALSE,
  points_shape = 21,
  points_size = 2,
  points_alpha = 0.5,
  points_jitter = TRUE,
  points_jitter_width = 0.1,
  points_gap = 0.1,
  bar_nudge = 0.075,
  bar_size = 1,
  bar_color = "black",
  line_color = "black",
  line_alpha = 0.5,
  font_size = 8,
  legend_position = "none",
  coord_flip = TRUE
)
}
\arguments{
\item{group_colors}{Default groups plotting colors}

\item{effect_colors}{Default effects plotting colors (i.e., Lower, Equivalent, Higher)}

\item{cloud_scale}{Scale of the cloud}

\item{cloud_alpha}{Cloud alpha/transparency}

\item{cloud_color}{Border color of the cloud}

\item{cloud_quantile_lines}{Should quantile lines be plotted? Default is TRUE}

\item{cloud_quantile_lines_size}{Size of the quantile lines}

\item{cloud_quantile_lines_color}{Color of the quantile lines}

\item{cloud_quantile_lines_adjust}{Should quantile lines be extended?}

\item{points_shape}{Point shape. Default is 21}

\item{points_size}{Point size}

\item{points_alpha}{Transparency of the points}

\item{points_jitter}{Should jitter be used?}

\item{points_jitter_width}{Width of the jitter}

\item{points_gap}{Gap between points and cloud}

\item{bar_nudge}{Gap between summary bar and cloud}

\item{bar_size}{Size of the summary bar}

\item{bar_color}{Color of the summary bar}

\item{line_color}{Line color. Default is "black"}

\item{line_alpha}{Line transparency. Default is 0.5}

\item{font_size}{Overall size of fonts}

\item{legend_position}{Position of the legend. Default is "none". See \code{\link[ggplot2]{theme}}
for more info}

\item{coord_flip}{Should figure be flipped?}
}
\value{
List with plot control elements
}
\description{
Returns default plotting parameters for \code{\link{plot_raincloud}},
   \code{\link{plot_raincloud_SESOI}}, and \code{\link{plot_pair_changes}}
}
\examples{
plot_raincloud(data.frame(data = rnorm(100)), value = "data")
plot_raincloud(
  iris,
  value = "Sepal.Length",
  groups = "Species",
  control = plot_control(points_size = 1)
)
}
