% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pair_BA.R
\name{plot_pair_BA}
\alias{plot_pair_BA}
\title{Bland-Altman plot}
\usage{
plot_pair_BA(
  predictor,
  outcome,
  SESOI_lower = 0,
  SESOI_upper = 0,
  confidence = 0.95,
  predictor_label = "Predictor",
  outcome_label = "Outcome",
  control = plot_control(),
  na.rm = FALSE
)
}
\arguments{
\item{predictor}{Numeric vector}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{confidence}{Default is 0.95}

\item{predictor_label}{Character vector. The name of the \code{predictor}. Default is "Predictor"}

\item{outcome_label}{Character vector. The name of the \code{outcome}. Default is "Outcome"}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Bland-Altman plot
}
\examples{
criterion <- rnorm(
  n = 100,
  mean = 100,
  sd = 10
)

practical <- criterion * 1.2 + rnorm(n = 100, mean = -12, sd = 5)

plot_pair_BA(practical,
  criterion,
  SESOI_lower = -10,
  SESOI_upper = 10,
  predictor_label = "Practical",
  outcome_label = "Criterion"
)
}
\seealso{
\code{\link{plot_pair_lm}} for lm residuals plot and
\code{\link{plot_pair_OLP}} for ordinary-least-squares residuals
plot
}
