% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pair_OLP.R
\name{plot_pair_OLP}
\alias{plot_pair_OLP}
\title{Ordinary Least Products plot}
\usage{
plot_pair_OLP(
  group_a,
  group_b,
  SESOI_lower = 0,
  SESOI_upper = 0,
  confidence = 0.95,
  control = plot_control(),
  na.rm = FALSE
)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, Post-test in the paired design, or "criterion" measure}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{confidence}{Default is 0.95}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
This function is used in reliability and validity visualization
}
\examples{
criterion <- rnorm(
  n = 100,
  mean = 100,
  sd = 10
)

practical <- criterion * 1.2 + rnorm(n = 100, mean = -12, sd = 5)

comment(practical) <- "practical"
comment(criterion) <- "criterion"

plot_pair_OLP(
  practical,
  criterion,
  SESOI_lower = -10,
  SESOI_upper = 10
)
}
\seealso{
\code{\link{plot_pair_BA}} for Bland-Altman plot and
   \code{\link{plot_pair_lm}} for lm residuals plot
}
