% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raincloud.R
\name{plot_raincloud}
\alias{plot_raincloud}
\title{Raincloud plot}
\usage{
plot_raincloud(
  data,
  value,
  value_label = value,
  groups = NULL,
  control = plot_control()
)
}
\arguments{
\item{data}{Data frame}

\item{value}{Character string. Name of  the column in \code{data}}

\item{value_label}{Character string. Label to be used. Default is \code{value}}

\item{groups}{Character string. Name of  the column in \code{data}. Default is \code{NULL}}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Raincloud plot
}
\examples{
plot_raincloud(data.frame(data = rnorm(100)), value = "data")
plot_raincloud(
  iris,
  value = "Sepal.Length",
  groups = "Species",
  control = plot_control(points_size = 1)
)
}
\seealso{
\code{\link{plot_raincloud_SESOI}}
}
