% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raincloud.R
\name{plot_raincloud_SESOI}
\alias{plot_raincloud_SESOI}
\title{Raincloud plot with smallest effect size of interest}
\usage{
plot_raincloud_SESOI(
  data,
  value,
  value_label = value,
  groups = NULL,
  SESOI_lower = 0,
  SESOI_upper = 0,
  control = plot_control()
)
}
\arguments{
\item{data}{Data frame}

\item{value}{Character string. Name of  the column in \code{data}}

\item{value_label}{Character string. Label to be used. Default is \code{value}}

\item{groups}{Character string. Name of  the column in \code{data}. Default is \code{NULL}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Raincloud plot with smallest effect size of interest
}
\examples{
plot_raincloud_SESOI(
  iris,
  value = "Sepal.Length",
  groups = "Species",
  SESOI_lower = 5,
  SESOI_upper = 7
)
}
\seealso{
\code{\link{plot_raincloud}}
}
