% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spaghetti.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Raincloud plot}
\usage{
plot_spaghetti(
  data,
  id,
  observations,
  observations_label = observations,
  value,
  value_label = value,
  SESOI_lower = 0,
  SESOI_upper = 0,
  control = plot_control()
)
}
\arguments{
\item{data}{Data frame}

\item{id}{Character string. Name of  the column in \code{data}}

\item{observations}{Character string. Name of the column in \code{data}}

\item{observations_label}{Character string. Label to be used for x-axis. Default is \code{observations}}

\item{value}{Character string. Name of  the column in \code{data}}

\item{value_label}{Character string. Label to be used. Default is \code{value}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Raincloud plot
}
\examples{
plot_spaghetti(
  sleep,
  id = "ID",
  observations = "group",
  value = "extra"
)
}
