% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPER.R
\name{PPER}
\alias{PPER}
\title{Proportion of Practically Equivalent Residuals}
\usage{
PPER(sigma, SESOI_lower = 0, SESOI_upper = 0, df = Inf)
}
\arguments{
\item{sigma}{Numeric}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{df}{Degrees of freedom. Used for t-distribution. Default is \code{Inf}
which gives normal distribution by default}
}
\value{
Numeric
}
\description{
Proportion of Practically Equivalent Residuals
}
\examples{
PPER(0.5, -1, 1)
}
