% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-predict.R
\name{predict.bmbstats_cv_model}
\alias{predict.bmbstats_cv_model}
\title{Predict from a \code{cv_model}}
\usage{
\method{predict}{bmbstats_cv_model}(object, new_data = NULL, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{cv_model} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{cv_model}
}
\examples{
m1 <- cv_model(
  Sepal.Length ~ . - Species,
  iris
)
predict(m1, new_data = iris)
}
