% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-predict-S3.R
\name{predict.bmbstats_RCT_predict}
\alias{predict.bmbstats_RCT_predict}
\title{S3 method for predicting \code{\link{RCT_predict}} results}
\usage{
\method{predict}{bmbstats_RCT_predict}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class "bmbstats_RCT_predict"}

\item{newdata}{Data used for prediction}

\item{...}{Extra arguments forwarded to \code{cv_model} predict}
}
\description{
S3 method for predicting \code{\link{RCT_predict}} results
}
\examples{
data("vertical_jump_data")

m1 <- cv_model(
  `Post-test` ~ `Pre-test` + Group,
  vertical_jump_data,
  control = model_control(
    cv_repeats = 10,
    cv_folds = 3,
    cv_strata = vertical_jump_data$Group
  )
)

m1_rct <- RCT_predict(
  m1,
  new_data = vertical_jump_data,
  outcome = "Post-test",
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control"
)

predict(m1_rct, vertical_jump_data)
}
