% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model-S3.R
\name{print.bmbstats_cv_model}
\alias{print.bmbstats_cv_model}
\title{S3 method for printing \code{\link{cv_model}} results}
\usage{
\method{print}{bmbstats_cv_model}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_cv_model}}

\item{...}{Extra arguments}
}
\description{
S3 method for printing \code{\link{cv_model}} results
}
\examples{
m1 <- cv_model(
  Sepal.Length ~ . - Species,
  iris
)
m1
}
