% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBI-S3.R
\name{print.bmbstats_MBI}
\alias{print.bmbstats_MBI}
\title{S3 method for printing \code{\link{bootstrap_MBI}} results}
\usage{
\method{print}{bmbstats_MBI}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_MBI}}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for printing \code{\link{bootstrap_MBI}} results
}
\examples{
mean_MBI <- bootstrap_MBI(
  describe_data(rnorm(10, 100, 20)),
  estimator = "mean",
  SESOI_lower = 95,
  SESOI_upper = 105
)
mean_MBI
plot(mean_MBI, control = plot_control(points_jitter = FALSE))
}
