% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MET-S3.R
\name{print.bmbstats_MET}
\alias{print.bmbstats_MET}
\title{S3 method for printing \code{\link{bootstrap_MET}} results}
\usage{
\method{print}{bmbstats_MET}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_MET}}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for printing \code{\link{bootstrap_MET}} results
}
\examples{
mean_MET <- bootstrap_MET(
  describe_data(rnorm(10, 100, 20)),
  estimator = "mean",
  SESOI_lower = 95,
  SESOI_upper = 105,
  alpha = 0.05
)
mean_MET
plot(mean_MET)
}
