% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations-MBI-S3.R
\name{print.bmbstats_observations_MBI}
\alias{print.bmbstats_observations_MBI}
\title{S3 method for printing \code{\link{observations_MBI}} results}
\usage{
\method{print}{bmbstats_observations_MBI}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_observations_MBI}}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for printing \code{\link{observations_MBI}} results
}
\examples{
data("bench_press_data")
obs_MBI <- observations_MBI(
  bench_press_data$Change,
  SESOI_lower = -5,
  SESOI_upper = 5,
  measurement_error = 2.5,
  confidence = 0.9
)

obs_MBI
plot(obs_MBI)
}
