% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations-MET-S3.R
\name{print.bmbstats_observations_MET}
\alias{print.bmbstats_observations_MET}
\title{S3 method for printing \code{\link{observations_MET}} results}
\usage{
\method{print}{bmbstats_observations_MET}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_observations_MET}}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for printing \code{\link{observations_MET}} results
}
\examples{
data("bench_press_data")
obs_MET <- observations_MET(
  bench_press_data$Change,
  SESOI_lower = -5,
  SESOI_upper = 5,
  measurement_error = 2.5,
  alpha = 0.05
)

obs_MET
plot(obs_MET)
}
