% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-analysis-S3.R
\name{print.bmbstats_RCT_analysis}
\alias{print.bmbstats_RCT_analysis}
\title{S3 method for printing \code{\link{RCT_analysis}} results}
\usage{
\method{print}{bmbstats_RCT_analysis}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_RCT_analysis}}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for printing \code{\link{RCT_analysis}} results
}
\examples{
set.seed(1666)

data("vertical_jump_data")

rct_model <- RCT_analysis(
  vertical_jump_data,
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control",
  pre_test = "Pre-test",
  post_test = "Post-test",
  control = model_control(boot_type = "perc", boot_samples = 500)
)

rct_model

plot(rct_model)
}
