% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-analysis.R
\name{RCT_analysis}
\alias{RCT_analysis}
\title{RCT Analysis}
\usage{
RCT_analysis(
  data,
  group,
  treatment_label,
  control_label,
  pre_test,
  post_test,
  SESOI_lower = SESOI_lower_RCT_func,
  SESOI_upper = SESOI_upper_RCT_func,
  estimator_function = RCT_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{group}{Character vector indicating the name of the column in \code{data} where group information is stored}

\item{treatment_label}{Character vector indicating the label inside the \code{group} column for the treatment group}

\item{control_label}{Character vector indicating the label inside the \code{group} column for the control group}

\item{pre_test}{Character vector indicating the name of the column in \code{data} where Pre-test observations are located}

\item{post_test}{Character vector indicating the name of the column in \code{data} where Post-test observations are located}

\item{SESOI_lower}{Function or numeric scalar. Default is \code{\link{SESOI_lower_RCT_func}}}

\item{SESOI_upper}{Function or numeric scalar. Default is \code{\link{SESOI_upper_RCT_func}}}

\item{estimator_function}{Function for providing RCT estimators. Default is \code{\link{RCT_estimators}}}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata}, and \code{confidence} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Object of class \code{bmbstats_RCT_analysis}
}
\description{
Wrapper function for \code{\link{bmbstats}} function to provide modular analysis for the Randomized Controlled Trials
}
\examples{
set.seed(1666)

data("vertical_jump_data")

rct_model <- RCT_analysis(
  vertical_jump_data,
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control",
  pre_test = "Pre-test",
  post_test = "Post-test",
  control = model_control(boot_type = "perc", boot_samples = 1000)
)

rct_model

plot(rct_model)
}
