% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-analysis.R
\name{RCT_estimators_simple}
\alias{RCT_estimators_simple}
\title{RCT Estimators - Simple}
\usage{
RCT_estimators_simple(
  control_pre_test,
  control_post_test,
  treatment_pre_test,
  treatment_post_test,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{control_pre_test}{Numeric vector containing Control Pre-test observations}

\item{control_post_test}{Numeric vector containing Control Post-test observations}

\item{treatment_pre_test}{Numeric vector containing Treatment Pre-test observations}

\item{treatment_post_test}{Numeric vector containing Treatment Post-test observations}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector
}
\description{
\code{RCT_estimators_simple} is used as a function for \code{\link{RCT_analysis}} function to provide a
simple (or reduced) estimators for RCT analysis
}
\examples{
set.seed(1666)

RCT_estimators_simple(
  control_pre_test = rnorm(20, 100, 10),
  control_post_test = rnorm(20, 105, 10),
  treatment_pre_test = rnorm(20, 100, 10),
  treatment_post_test = rnorm(20, 120, 10),
  SESOI_lower = -5,
  SESOI_upper = 5
)
}
