% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct-predict.R
\name{RCT_predict}
\alias{RCT_predict}
\title{RCT Predict}
\usage{
RCT_predict(
  model,
  new_data,
  outcome,
  group,
  control_label,
  treatment_label,
  subject_label = rownames(new_data),
  na.rm = FALSE
)
}
\arguments{
\item{model}{Object returned by \code{\link{cv_model}} function}

\item{new_data}{Data frame}

\item{outcome}{Character string indicating the outcome column in \code{new_data}}

\item{group}{Character vector indicating the name of the column in \code{data} where group information is stored}

\item{control_label}{Character vector indicating the label inside the \code{group} column for the control group}

\item{treatment_label}{Character vector indicating the label inside the \code{group} column for the treatment group}

\item{subject_label}{Row labels, usually participants names. Default is \code{new_data} row names}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Object of class \code{bmbstats_RCT_predict}
}
\description{
\code{RCT_predict} perform additional analysis on \code{model} returned by \code{\link{cv_model}} function
assuming RCT data
}
\examples{
data("vertical_jump_data")

m1 <- cv_model(
  `Post-test` ~ `Pre-test` + Group,
  vertical_jump_data,
  control = model_control(
    cv_repeats = 10,
    cv_folds = 3,
    cv_strata = vertical_jump_data$Group
  )
)

m1_rct <- RCT_predict(
  m1,
  new_data = vertical_jump_data,
  outcome = "Post-test",
  group = "Group",
  treatment_label = "Treatment",
  control_label = "Control"
)

m1_rct
plot(m1_rct)
}
