% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_relationship.R
\name{relationship_lm_estimators}
\alias{relationship_lm_estimators}
\title{Dependent groups linear relationship estimators}
\usage{
relationship_lm_estimators(
  predictor,
  outcome,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{predictor}{Numeric vector}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named numeric vector with estimators
}
\description{
Function that provides a list of estimators for linear relationship between \code{predictor} and \code{outcome}.
User is free to define his own list of estimators. Used in \code{\link{compare_dependent_groups}}
}
\examples{
data("yoyo_mas_data")
predictor <- yoyo_mas_data$YoYoIR1
outcome <- yoyo_mas_data$MAS

relationship_lm_estimators(predictor, outcome, SESOI_lower = -0.5, SESOI_upper = 0.5)
}
