% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_analysis.R
\name{reliability_estimators}
\alias{reliability_estimators}
\title{Reliability estimators}
\usage{
reliability_estimators(
  data,
  trial1,
  trial2,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{trial1}{Character vector indicating column name in the \code{data}}

\item{trial2}{Character vector indicating column name in the \code{data}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named vector with reliability estimators
}
\description{
\code{reliability_estimators} provides reliability estimators using the OLP regression model, where
\code{trial1} is the outcome variable, and \code{trial2} is the predictor.
}
\examples{
data("agreement_data")

reliability_estimators(
  data = agreement_data,
  trial1 = "Practical_score.trial1",
  trial2 = "Practical_score.trial2"
)
}
