% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dependent_groups.R
\name{SESOI_lower_dependent_func}
\alias{SESOI_lower_dependent_func}
\title{SESOI lower threshold for \code{\link{compare_dependent_groups}}}
\usage{
SESOI_lower_dependent_func(pre, post, na.rm = FALSE)
}
\arguments{
\item{pre}{Numeric vector}

\item{post}{Numeric vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
SD of \code{pre} multiplied by 0.2 (Cohen's trivial)
}
\description{
SESOI lower threshold for \code{\link{compare_dependent_groups}}
}
\examples{
SESOI_lower_dependent_func(rnorm(20), rnorm(10))
}
