% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model-fit.R
\name{SESOI_lower_func}
\alias{SESOI_lower_func}
\title{SESOI lower function}
\usage{
SESOI_lower_func(predictors, outcome, na.rm = FALSE)
}
\arguments{
\item{predictors}{Data frame}

\item{outcome}{Vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
Function that calculates lower SESOI threshold
}
\examples{
SESOI_lower_func(outcome = iris[[1]])
}
