% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_analysis.R
\name{SESOI_lower_validity_func}
\alias{SESOI_lower_validity_func}
\title{SESOI lower threshold for \code{\link{validity_analysis}}}
\usage{
SESOI_lower_validity_func(data, criterion, practical, na.rm = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{criterion}{Character vector. Column name in \code{data} data frame}

\item{practical}{Character vector. Column name in \code{data} data frame}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
SD of \code{criterion} column of the \code{data}
data frame multiplied by -0.2 (Cohen's trivial)
}
\description{
SESOI lower threshold for \code{\link{validity_analysis}}
}
\examples{
data("agreement_data")
SESOI_lower_validity_func(
  data = agreement_data,
  criterion =  "Criterion_score.trial1",
  practical = "Practical_score.trial1"
)
}
