% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-fit.R
\name{SESOI_upper_func}
\alias{SESOI_upper_func}
\title{SESOI upper function}
\usage{
SESOI_upper_func(predictors, outcome, na.rm = FALSE)
}
\arguments{
\item{predictors}{Data frame}

\item{outcome}{Vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\description{
Function that calculates upper SESOI threshold
}
\examples{
SESOI_upper_func(outcome = iris[[1]])
}
