% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_relationship.R
\name{SESOI_upper_relationship_func}
\alias{SESOI_upper_relationship_func}
\title{SESOI upper threshold for \code{\link{compare_dependent_groups}}}
\usage{
SESOI_upper_relationship_func(predictor, outcome, na.rm = FALSE)
}
\arguments{
\item{predictor}{Numeric vector}

\item{outcome}{Vector}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
SD of \code{outcome} multiplied by 0.2 (Cohen's trivial)
}
\description{
SESOI upper threshold for \code{\link{compare_dependent_groups}}
}
\examples{
SESOI_upper_relationship_func(rnorm(20), rnorm(10))
}
