% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_analysis.R
\name{SESOI_upper_reliability_func}
\alias{SESOI_upper_reliability_func}
\title{SESOI upper threshold for \code{\link{reliability_analysis}}}
\usage{
SESOI_upper_reliability_func(data, trial1, trial2, na.rm = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{trial1}{Character vector. Column name in \code{data} data frame}

\item{trial2}{Character vector. Column name in \code{data} data frame}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Pooled SD of \code{trial1} and \code{trial2} columns of the \code{data}
data frame multiplied by 0.2 (Cohen's trivial)
}
\description{
SESOI upper threshold for \code{\link{reliability_analysis}}
}
\examples{
data("agreement_data")
SESOI_lower_reliability_func(
  data = agreement_data,
  trial1 = "Practical_score.trial1",
  trial2 = "Practical_score.trial2"
)
}
