% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_analysis.R
\name{validity_analysis}
\alias{validity_analysis}
\title{Validity Analysis}
\usage{
validity_analysis(
  data,
  criterion,
  practical,
  SESOI_lower = SESOI_lower_validity_func,
  SESOI_upper = SESOI_upper_validity_func,
  estimator_function = validity_estimators,
  control = model_control(),
  na.rm = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{criterion}{Character vector indicating column name in the \code{data}}

\item{practical}{Character vector indicating column name(s) in the \code{data}}

\item{SESOI_lower}{Function or numeric scalar. Default is \code{\link{SESOI_lower_validity_func}}}

\item{SESOI_upper}{Function or numeric scalar. Default is \code{\link{SESOI_upper_validity_func}}}

\item{estimator_function}{Function for providing validity estimators. Default is \code{\link{validity_estimators}}}

\item{control}{Control object returned from \code{\link{model_control}} function.
Use \code{boot_type}, \code{boot_samples}, \code{boot_strata}, and \code{confidence} to setup bootstrap.}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Object of class \code{bmbstats}
}
\description{
\code{validity_analysis} represents a wrapper function for the \code{\link{bmbstats}} function.
\code{validity_analysis} runs the bootstrap validity analysis of the \code{data} data frame using
\code{estimator_function} to return the estimators
}
\examples{
data("agreement_data")

val_analysis <- validity_analysis(
  data = agreement_data,
  criterion = "Criterion_score.trial1",
  practical = "Practical_score.trial1",
  control = model_control(
    boot_type = "perc",
    boot_samples = 1000,
    seed = 1667)
)

val_analysis

plot(val_analysis)
}
