% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_analysis.R
\name{validity_estimators}
\alias{validity_estimators}
\title{Validity estimators}
\usage{
validity_estimators(
  data,
  criterion,
  practical,
  SESOI_lower = 0,
  SESOI_upper = 0,
  na.rm = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{criterion}{Character vector indicating column name in the \code{data}}

\item{practical}{Character vector indicating column name in the \code{data}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Named vector with validity estimators
}
\description{
\code{validity_estimators} provides validity estimators using the OLP regression model, where
\code{criterion} is the outcome variable, and \code{practical} is the predictor.
}
\examples{
data("agreement_data")

validity_estimators(
  data = agreement_data,
  criterion = "Criterion_score.trial1",
  practical = "Practical_score.trial1"
)
}
