% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_arguments.R
\name{basic_arguments}
\alias{basic_arguments}
\title{Basic arguments}
\arguments{
\item{x}{Numeric vector}

\item{model}{Model object}

\item{predictors}{Data frame}

\item{outcome}{Vector}

\item{observed}{Numeric vector}

\item{predicted}{Numeric vector}

\item{group_a}{Numeric vector. This group represents baseline/control, observed variable, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, predicted variable, Post-test in the paired design, or "criterion" measure}

\item{group_a_label}{Character vector. The name of the \code{group_a}. Default is "Group A"}

\item{group_b_label}{Character vector. The name of the \code{group_b}. Default is "Group B"}

\item{fitted_label}{Character vector. The label to be used for fitted. Default is "Fitted"}

\item{residuals_label}{Character vector. The label to be used for residuals. Default is "Residuals"}

\item{paired}{Paired groups? Default is \code{FALSE}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{confidence}{Default is 0.95}
}
\description{
Basic arguments
}
