% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats-package.R
\docType{package}
\name{bmbstats-package}
\alias{bmbstats-package}
\alias{_PACKAGE}
\title{bmbstats: Bootstrap Magnitude-based Statistics}
\description{
Set of functions for analysis and graphing of common sport science problems.
}
\author{
\strong{Maintainer}: Mladen Jovanović \email{coach.mladen.jovanovic@gmail.com}

}
\keyword{internal}
