% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbstats-package.R
\docType{package}
\name{bmbstats-package}
\alias{bmbstats-package}
\alias{_PACKAGE}
\title{bmbstats: Bootstrap Magnitude-based Statistics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Set of functions for analysis and graphing of common sport science problems. Statistical
    inference is done using bootstrap method. Extensive documentation is available in the links provided,
    particularly in the "bmbstats: Bootstrap Magnitude-based Statistics for Sports Scientists" book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mladenjovanovic.github.io/bmbstats/}
  \item \url{https://github.com/mladenjovanovic/bmbstats}
  \item \url{https://mladenjovanovic.github.io/bmbstats-book/}
  \item \url{https://github.com/mladenjovanovic/bmbstats-book}
  \item Report bugs at \url{https://github.com/mladenjovanovic/bmbstats/issues}
}

}
\author{
\strong{Maintainer}: Mladen Jovanović \email{coach.mladen.jovanovic@gmail.com}

}
\keyword{internal}
