% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLES.R
\name{CLES}
\alias{CLES}
\title{Common Language Effect Size (CLES)}
\usage{
CLES(
  group_a,
  group_b,
  SESOI_lower = 0,
  SESOI_upper = 0,
  method = "algebraic",
  count_equivalent = "ignore",
  na.rm = FALSE
)
}
\arguments{
\item{group_a}{Vector of observations from Group A. This groups represents baseline/control}

\item{group_b}{Vector of observations from Group B.}

\item{SESOI_lower}{Smallest Effect Size Of Interest. Uses to differentiate between lower/equivalent difference. Default is 0}

\item{SESOI_upper}{Smallest Effect Size Of Interest. Uses to differentiate between equivalent/higher difference. Default is 0}

\item{method}{Select "brute-force" or "algebraic" method to calculate CLES. Default is "algebraic"}

\item{count_equivalent}{How should equivalent differences be counted? Ignored, counted as higher or lower difference.
Default is "ignore". Other options are "higher" and "lower"}

\item{na.rm}{Should NAs be removed? Default is FALSE}
}
\value{
Numeric
}
\description{
\code{CLES} calculates the probability that a random observation from Group B is higher/larger
than a random observation from Group A. Groups are considered independent.
}
\examples{
CLES(
  rnorm(n = 50, mean = 100, sd = 10),
  rnorm(n = 40, mean = 98, sd = 8)
)
}
