% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_control.R
\name{model_control}
\alias{model_control}
\title{Model Control Constructor}
\usage{
model_control(seed = round(stats::runif(1, 1, 10000)))
}
\arguments{
\item{seed}{Random number seed}
}
\description{
Model Control Constructor
}
