% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-control.R
\name{model_control}
\alias{model_control}
\title{Model Control Constructor}
\usage{
model_control(
  cv_folds = 5,
  cv_repeats = 10,
  cv_strata = NULL,
  boot_type = "bca",
  boot_samples = 2000,
  boot_strata = NULL,
  confidence = 0.95,
  iter = TRUE,
  seed = round(stats::runif(1, 1, 10000))
)
}
\arguments{
\item{cv_folds}{Number of cross-validation folds. Default is 5}

\item{cv_repeats}{Number of cross-validation repeats. Default is 10}

\item{cv_strata}{Strata to be used for cross-validation. Default is \code{NULL}}

\item{boot_type}{Type of bootstrap confidence interval ("bca" and "perc"). Default is "bca"}

\item{boot_samples}{Number of bootstrap resamples. Default is 2000}

\item{boot_strata}{Strata to be used for bootstrap. Default is \code{NULL}}

\item{confidence}{Confidence to be used for calculating bootstrap CIs. Default is 0.95}

\item{iter}{Should progress bar be shown? Default is \code{TRUE}}

\item{seed}{Random number seed}
}
\description{
Model Control Constructor
}
