% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLP.R
\name{OLP_regression}
\alias{OLP_regression}
\title{OLP regression}
\source{
Function modified using the function from
\url{https://stat.ethz.ch/pipermail/r-help/2011-July/285022.html}
}
\usage{
OLP_regression(group_a, group_b, na.rm = FALSE)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, observed variable, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, predicted variable, Post-test in the paired design, or "criterion" measure}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
List with the following elements: \code{intercept}, \code{slope}, \code{rse}
}
\description{
OLP regression
}
\examples{
data("weight_data")
with(
  weight_data,
  OLP_regression(`OS 1`, `OS 3`)
)
}
