% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model-S3.R
\name{plot.bmbstats_cv_model}
\alias{plot.bmbstats_cv_model}
\title{S3 method for plotting \code{\link{cv_model}} results}
\usage{
\method{plot}{bmbstats_cv_model}(x, type = "residuals", ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_cv_model}}

\item{type}{Type of plot. Options are "residuals",  being default}

\item{...}{Extra arguments}
}
\description{
S3 method for plotting \code{\link{cv_model}} results
}
\examples{
m1 <- cv_model(
  Sepal.Length~. -Species,
  iris
)
plot(m1, "residuals")
}
