% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-S3.R
\name{plot.bmbstats_cv_model}
\alias{plot.bmbstats_cv_model}
\title{S3 method for plotting \code{\link{cv_model}} results}
\usage{
\method{plot}{bmbstats_cv_model}(x, type = "residuals", ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_cv_model}}

\item{type}{Type of plot. Options are "residuals", "training-residuals",
"testing-residuals", "bias-variance-index", "bias-variance-observed",
"prediction-index", "prediction-observed", and "estimators".  Default is "residuals"}

\item{...}{Extra arguments. Use \code{\link{plot_control}} to control plotting style}
}
\description{
S3 method for plotting \code{\link{cv_model}} results
}
\examples{
m1 <- cv_model(
  Sepal.Length ~ . - Species,
  iris
)
plot(m1, "residuals")
}
