% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-control.R
\name{plot_control}
\alias{plot_control}
\title{Plot control constructor}
\usage{
plot_control(
  group_colors = c("#5DA5DA", "#F15854", "#4D4D4D", "#60BD68", "#FAA43A", "#F17CB0",
    "#B276B2", "#DECF3F"),
  effect_colors = c(Lower = "#F158546A", Equivalent = "#4D4D4D6A", Higher =
    "#60BD686A"),
  cloud_scale = 0.5,
  cloud_alpha = 0.6,
  cloud_color = NA,
  cloud_quantile_lines = TRUE,
  cloud_quantile_lines_size = 0.4,
  cloud_quantile_lines_color = "white",
  cloud_quantile_lines_adjust = FALSE,
  points_shape = 21,
  points_color = "black",
  points_fill = "black",
  points_size = 2,
  points_alpha = 0.5,
  points_jitter = TRUE,
  points_jitter_width = 0.1,
  points_gap = 0.1,
  summary_bar_nudge = 0.075,
  summary_bar_size = 1,
  summary_bar_color = "black",
  summary_bar_alpha = 0.8,
  bar_size = 3,
  bar_alpha = 0.5,
  line_size = 1,
  line_color = "black",
  line_alpha = 0.5,
  identity_line = TRUE,
  SESOI_band = TRUE,
  SESOI_color = "#4D4D4D6A",
  SESOI_alpha = 0.2,
  font_size = 8,
  legend_position = "none",
  smooth_method = "lm",
  smooth_se = FALSE,
  smooth_color = "#5DA5DA",
  smooth_fill = "#4D4D4D6A",
  smooth_alpha = 0.8,
  smooth_size = 1,
  bins = 30,
  panel_labels = c("A", "B", "C", "D", "E", "F"),
  panel_labels_size = 10,
  text_label_size = 3,
  coord_flip = TRUE,
  sort = TRUE
)
}
\arguments{
\item{group_colors}{Default groups plotting colors}

\item{effect_colors}{Default effects plotting colors (i.e., Lower, Equivalent, Higher)}

\item{cloud_scale}{Scale of the cloud}

\item{cloud_alpha}{Cloud alpha/transparency}

\item{cloud_color}{Border color of the cloud}

\item{cloud_quantile_lines}{Should quantile lines be plotted? Default is TRUE}

\item{cloud_quantile_lines_size}{Size of the quantile lines}

\item{cloud_quantile_lines_color}{Color of the quantile lines}

\item{cloud_quantile_lines_adjust}{Should quantile lines be extended?}

\item{points_shape}{Point shape. Default is 21}

\item{points_color}{Point color. Default is "black"}

\item{points_fill}{Point fill. Default is "transparent"}

\item{points_size}{Point size}

\item{points_alpha}{Transparency of the points}

\item{points_jitter}{Should jitter be used?}

\item{points_jitter_width}{Width of the jitter}

\item{points_gap}{Gap between points and cloud}

\item{summary_bar_nudge}{Gap between summary bar and cloud}

\item{summary_bar_size}{Size of the summary bar}

\item{summary_bar_color}{Color of the summary bar}

\item{summary_bar_alpha}{Transparency of the summary bar}

\item{bar_size}{Size of the bar}

\item{bar_alpha}{Transparency of the bar}

\item{line_size}{Line size}

\item{line_color}{Line color. Default is "black"}

\item{line_alpha}{Line transparency. Default is 0.5}

\item{identity_line}{Should identity line be plotted?}

\item{SESOI_band}{Should SESOI band be plotted?}

\item{SESOI_color}{Which color should \code{SESOI_band} use?}

\item{SESOI_alpha}{Transparency for the SESOI band. Default is 0.2}

\item{font_size}{Overall size of fonts}

\item{legend_position}{Position of the legend. Default is "none". See \code{\link[ggplot2]{theme}}
for more info}

\item{smooth_method}{Smoothing method (function) to use, accepts either a character vector, e.g.
"auto", "lm", "glm", "gam", "loess" or a function, e.g. MASS::rlm or mgcv::gam, stats::lm,
or stats::loess. Default is "loess". See \code{\link[ggplot2]{geom_smooth}} for more info}

\item{smooth_se}{Should standard error be plotted? Defaults is \code{FALSE}}

\item{smooth_color}{Color of the smooth line}

\item{smooth_fill}{Color of the smooth confidence area}

\item{smooth_alpha}{Transparency for the smooth line}

\item{smooth_size}{Size of the smooth line}

\item{bins}{Number of bins in histogram. Default is 30}

\item{panel_labels}{Labels for panels}

\item{panel_labels_size}{Size of panel labels. Default is 10}

\item{text_label_size}{Size of the text labels. Default is 3}

\item{coord_flip}{Should figure be flipped?}

\item{sort}{Should items be sorted? Default is TRUE}
}
\value{
List with plot control elements
}
\description{
Returns default plotting parameters for \code{\link{plot_raincloud}},
\code{\link{plot_raincloud_SESOI}}, and \code{\link{plot_pair_changes}}
}
\examples{
plot_raincloud(data.frame(data = rnorm(100)), value = "data")
plot_raincloud(
  iris,
  value = "Sepal.Length",
  groups = "Species",
  control = plot_control(points_size = 1)
)
}
