% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pair_BA.R
\name{plot_pair_BA}
\alias{plot_pair_BA}
\title{Bland-Altman plot}
\usage{
plot_pair_BA(
  group_a,
  group_b,
  SESOI_lower = 0,
  SESOI_upper = 0,
  confidence = 0.95,
  group_a_label = "Group A",
  group_b_label = "Group B",
  control = plot_control(),
  na.rm = FALSE
)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, Post-test in the paired design, or "criterion" measure}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{confidence}{Default is 0.95}

\item{group_a_label}{Character vector. The name of the \code{group_a}. Default is "Group A"}

\item{group_b_label}{Character vector. The name of the \code{group_b}. Default is "Group B"}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Bland-Altman plot
}
\examples{
criterion <- rnorm(
  n = 100,
  mean = 100,
  sd = 10
)

practical <- criterion * 1.2 + rnorm(n = 100, mean = -12, sd = 5)

plot_pair_BA(practical,
  criterion,
  SESOI_lower = -10,
  SESOI_upper = 10,
  group_a_label = "Practical",
  group_b_label = "Criterion"
)
}
\seealso{
\code{\link{plot_pair_lm}} for lm residuals plot and
\code{\link{plot_pair_OLP}} for ordinary-least-squares residuals
plot
}
