% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pair-OLP.R
\name{plot_pair_OLP}
\alias{plot_pair_OLP}
\title{Ordinary Least Products plot}
\usage{
plot_pair_OLP(
  predictor,
  outcome,
  SESOI_lower = 0,
  SESOI_upper = 0,
  confidence = 0.95,
  predictor_label = "Predictor",
  outcome_label = "Outcome",
  fitted_label = "Fitted",
  residuals_label = "Residuals",
  control = plot_control(),
  na.rm = FALSE
)
}
\arguments{
\item{predictor}{Numeric vector}

\item{outcome}{Vector}

\item{SESOI_lower}{Lower smallest effect size of interest threshold}

\item{SESOI_upper}{Upper smallest effect size of interest threshold}

\item{confidence}{Default is 0.95}

\item{predictor_label}{Character vector. The name of the \code{predictor}. Default is "Predictor"}

\item{outcome_label}{Character vector. The name of the \code{outcome}. Default is "Outcome"}

\item{fitted_label}{Character vector. The label to be used for fitted. Default is "Fitted"}

\item{residuals_label}{Character vector. The label to be used for residuals. Default is "Residuals"}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
This function is used in reliability and validity visualization
}
\examples{
criterion <- rnorm(
  n = 100,
  mean = 100,
  sd = 10
)

practical <- criterion * 1.2 + rnorm(n = 100, mean = -12, sd = 5)

plot_pair_OLP(practical,
  criterion,
  SESOI_lower = -10,
  SESOI_upper = 10,
  predictor_label = "Practical",
  outcome_label = "Criterion"
)
}
\seealso{
\code{\link{plot_pair_BA}} for Bland-Altman plot and
\code{\link{plot_pair_lm}} for lm residuals plot
}
