% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Spaghetti plot}
\usage{
plot_spaghetti(
  data,
  id,
  observations,
  observations_label = observations,
  value,
  value_label = value,
  SESOI_lower = 0,
  SESOI_upper = 0,
  control = plot_control()
)
}
\arguments{
\item{data}{Data frame}

\item{id}{Character string. Name of  the column in \code{data}}

\item{observations}{Character string. Name of the column in \code{data}}

\item{observations_label}{Character string. Label to be used for x-axis. Default is \code{observations}}

\item{value}{Character string. Name of  the column in \code{data}}

\item{value_label}{Character string. Label to be used. Default is \code{value}}

\item{SESOI_lower}{Lower smallest effect size of interest threshold. Default is 0}

\item{SESOI_upper}{Upper smallest effect size of interest threshold.  Default is 0}

\item{control}{Plotting control object returned from \code{\link{plot_control}}}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Spaghetti plot
}
\examples{
test_data <- expand.grid(
  id = 1:10,
  obs = 1:10
)

test_data$val <- with(
  test_data,
  rnorm(nrow(test_data), obs, id)
)

plot_spaghetti(
  test_data,
  id = "id",
  observations = "obs",
  value = "val",
  SESOI_lower = -1,
  SESOI_upper = 1,
  control = plot_control(
    points_shape = 21,
    points_fill = "white",
    points_alpha = 1,
    points_size = 0.75,
    line_size = 0.75,
    legend_position = "right"
  )
)
}
