% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-model-S3.R
\name{print.bmbstats_cv_model}
\alias{print.bmbstats_cv_model}
\title{S3 method for printing \code{\link{cv_model}} results}
\usage{
\method{print}{bmbstats_cv_model}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bmbstats_cv_model}}

\item{...}{Extra arguments}
}
\description{
S3 method for printing \code{\link{cv_model}} results
}
\examples{
data("vertical_jump_data")

m1 <- cv_model(
  `Post-test` ~ `Pre-test` * Group * `Squat 1RM`,
  vertical_jump_data,
  control = model_control(
    cv_repeats = 10,
    cv_folds = 3,
    cv_strata = vertical_jump_data$Group
  )
)

m1
plot(m1, "residuals")
}
