% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Calculates pooled SD}
\usage{
sd_pooled(group_a, group_b, na.rm = FALSE)
}
\arguments{
\item{group_a}{Numeric vector. This group represents baseline/control, Pre-test in the paired design, or "practical" measure}

\item{group_b}{Numeric vector. This group represents experimental, Post-test in the paired design, or "criterion" measure}

\item{na.rm}{Should NAs be removed? Default is \code{FALSE}}
}
\value{
Pooled SD
}
\description{
\code{sd_pooled} calculates pooled standard deviation from vectors \code{group_a} and \code{group_b}
}
